requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==3.2.25
django-cloudinary-storage==0.3.0
djangorestframework==3.12.4
idna==3.11
pillow==12.0.0
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
